refactor(generator): Remove the generator package
The generator package provided limited functionality used by two very specific modules. The generator is no longer relevant in our current architecture, so we've moved its functionality inline to the methods that need it.
